MODUL-1181 - retirer fichier de classe inutile et régler typo
